+++ src/Http/ApiResponseData.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Dto;
+namespace App\Http;
 
 use OpenApi\Annotations as OA;
